+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master (unreleased)
 
+- Fix running background data migrations inline
+
 ## 0.31.1 (2025-11-11)
 
 - Do not attempt running background migrations on database with `database_tasks: false`